+++ project/source_code/drivetrain/src/engine_simulator.cpp
@@ -72,7 +72,7 @@ void Engine::RPM(const uint8_t &acc_pedal, const uint8_t &brk_pedal, const uint1
     * @param speed current vehicle speed.
     * @param speed_to_rpm_factor factor to multiply with speed to receive actual RPM.
 */
-void Engine::ActualRPM(const float &speed, const double &speed_to_rpm_factor)
+void Engine::ActualRPM(const float &speed, const float &speed_to_rpm_factor)
 {
     if(this->eng_sts)
     {
